requirements.txt
# requirements.txt

python-telegram-bot==20.3
telethon==1.31.0
apify-client==0.0.10
python-dotenv==1.0.0


#pip install -r requirements.txt
